pyproject.toml
[project]
name = "autotrain"
version = "0.1.0"
description = "Autonomous ML training platform — train, evaluate, self-correct, repeat."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.50.0",
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "requests>=2.31",
    "streamlit>=1.30",
    "plotly>=5.18",
]

[project.scripts]
autotrain = "autotrain.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autotrain"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.9",
]
